// modules/transfer/modules/bulk-transfer/enum/bt-view.enum.ts
export enum EBTTypeTransaction {
  DEFAULT = 'default',
  SIGNATURE_TRACKING = 'signatureTracking',
  SIGNATURE_TRACKING_DEFAULT = 'signatureTrackingDefault',
  SIGNATURE_TRACKING_OPERATION = 'signatureTrackingOperation'
}

export enum EBViewMode {
  HISTORY_TRANSACTION = 'historyTransaction',
  SIGNATURE_TRACKING = 'signatureTracking',
  SIGNATURE_TRACKING_OPERATION = 'signatureTrackingOperation',
  DEFAULT = 'default'
}

export enum EBTScheduleService {
  BULK_TRANSFER = 'ach-cararc'
}