// modules/transfer/modules/bulk-transfer/enum/bt-navigation.enum.ts
export enum EBTUrlNavigateCollection {
  HOME = '/transfer/bulk-transfers',
  DEFAULT_CONFIRMATION = '/transfer/bulk-transfers/confirmation',
  DEFAULT_VOUCHER = '/transfer/bulk-transfers/voucher',
  SIGNATURE_TRACKING_OPERATIONS = '/transfer/bulk-transfers/st-operation',
  SIGNATURE_TRACKING_VOUCHER = '/transfer/bulk-transfers/st-voucher',
  SIGNATURE_TRACKING_MODIFY = '/transfer/bulk-transfers/stm-home',
  TRANSACTION_HISTORY_VOUCHER = '/transfer/bulk-transfers/th-voucher',
}